+++ Sources/MapboxSearchUI/CategoryDataProvider.swift
@@ -5,8 +5,8 @@ public protocol CategoryDataProvider {
     /// Provide custom list of category buttons in the row under search textfield.
     /// Only first 4 categories would be used.
     /// Default values would be added if necessary.
-    var categorySlots: [Category] { get }
+    var categorySlots: [SearchCategory] { get }
     
     /// Provide custom vertical list of categories
-    var categoryList: [Category] { get }
+    var categoryList: [SearchCategory] { get }
 }